+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## 0.0.2 - 2014-03-02
+- add suite name completion
+- add options completion
+
 ## 0.0.1 - 2014-02-13
 - codecept alias and command completion